requirements.txt
# Ghost SDR v3.1 — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core Framework ────────────────────────────────────────────────
fastapi==0.128.8
uvicorn==0.39.0
python-dotenv==1.2.1
pydantic==2.12.5

# ── Database ──────────────────────────────────────────────────────
supabase==2.28.0

# ── HTTP & Signal Collection ──────────────────────────────────────
requests==2.32.5
httpx==0.28.1
youtube-transcript-api==0.6.2
python-multipart==0.0.20
typing-extensions==4.15.0

# ── Free LLM Providers ────────────────────────────────────────────
# Switch between them via MODEL_PROVIDER in .env — install all for flexibility

# Gemini 1.5 Flash — RECOMMENDED (free, 1,500 req/day)
# Get key: https://aistudio.google.com/app/apikey
google-generativeai==0.8.3

# Groq Llama 3.3 70B — fastest free option (14,400 req/day)
# Get key: https://console.groq.com
groq==1.0.0

# Ollama — Local LLM (free)
ollama==0.6.1

# ── Paid Upgrade Path ─────────────────────────────────────────────
# Keep even on free tier for quality comparison
anthropic==0.84.0
